Move scripts from public to src directory.
