+++ frontend/src/scripts/typeWriter.js
